Refactors StorefrontInterfaceProvider UI methods to StorefrontInterfaceProvider file
